+++ src/Plugins/Flight/Import/CsvImport/src/CsvImportSettings.h
@@ -45,8 +45,8 @@ public:
      * These values are peristed in the application settings.
      */
     enum struct Format {
-        SkyDolly = 0,
-        FlightRadar24 = 1,
+        // Removed: SkyDolly = 0,
+        Flightradar24 = 1,
         FlightRecorder = 2
     };